refactor: CLI credential, presentation
